// src/lib/types.ts
export interface SerializedShape extends SerializedBase {
  type: 'shape'
  widthM: number
  heightM: number
  color: string
  baseWidthPx: number
  baseHeightPx: number
  width: number
  height: number
}

export interface SerializedLine extends SerializedBase {
  type: 'line'
  x1: number
  y1: number
  x2: number
  y2: number
  lengthM: number
  color: string
  strokeWidth: number
}

export interface SerializedMask extends SerializedBase {
  type: 'mask'
  width: number
  height: number
}

export interface SerializedImage extends SerializedBase {
  type: 'backgroundImage' | 'overlayImage'
  imageData: string
  originX: string
  originY: string
}

export type SerializedObject =
  | SerializedShape
  | SerializedLine
  | SerializedMask
  | SerializedImage

export interface SerializedProject {
  version: number
  pixelsPerMeter: number | null
  backgroundImage: string | null
  savedAt: string
  objects: SerializedObject[]
  id?: string // IndexedDB key
}